+++ src/replaceAll.js
@@ -1,9 +1,9 @@
 import { curryN, invoker } from 'ramda';
 
 import isFunction from './isFunction';
-import polyfill from './internal/polyfills/String.replaceAll';
+import ponyfill from './internal/ponyfills/String.replaceAll';
 
-export const replaceAllPolyfill = curryN(3, polyfill);
+export const replaceAllPonyfill = curryN(3, ponyfill);
 
 export const replaceAllInvoker = invoker(2, 'replaceAll');
 
@@ -28,9 +28,8 @@ export const replaceAllInvoker = invoker(2, 'replaceAll');
  * RA.replaceAll(/x/g, 'v', 'xxx'); //=> 'vvv'
  * RA.replaceAll(/x/, 'v', 'xxx'); //=> TypeError
  */
-
 const replaceAll = isFunction(String.prototype.replaceAll)
   ? replaceAllInvoker
-  : replaceAllPolyfill;
+  : replaceAllPonyfill;
 
 export default replaceAll;