+++ test/replaceAll.js
@@ -1,7 +1,7 @@
 import { assert } from 'chai';
 
 import * as RA from '../src';
-import { replaceAllPolyfill, replaceAllInvoker } from '../src/replaceAll';
+import { replaceAllPonyfill, replaceAllInvoker } from '../src/replaceAll';
 
 describe('replaceAll', function () {
   it('should replace all matches', function () {
@@ -36,10 +36,12 @@ describe('replaceAll', function () {
     assert.strictEqual(actual, expected);
   });
 
-  context('given searchValue is a non-global RegExp', function () {
-    specify('should throw Error', function () {
-      assert.throws(() => RA.replaceAll(/a/, 'c', 'abc'), TypeError);
-    });
+  it('should properly process special characters', function () {
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace#Specifying_a_string_as_a_parameter
+    // https://github.com/zloirock/core-js/issues/900
+    const actual = RA.replaceAll('.', '$$$$$', 'foo.bar');
+    const expected = 'foo.bar'.replaceAll('.', '$$$$$');
+    assert.strictEqual(actual, expected);
   });
 
   it('should be curried', function () {
@@ -118,16 +120,10 @@ describe('replaceAll', function () {
     });
   });
 
-  context('replaceAllPolyfill', function () {
-    context('given searchValue is a non-global RegExp', function () {
-      specify('should throw Error', function () {
-        assert.throws(() => replaceAllPolyfill(/a/, 'c', 'abc'), TypeError);
-      });
-    });
-
+  context('replaceAllPonyfill', function () {
     specify('should support global RegExp searchValue', function () {
       const value = 'xxx';
-      const actual = replaceAllPolyfill(/x/g, 'v', value);
+      const actual = replaceAllPonyfill(/x/g, 'v', value);
       const expected = 'vvv';
 
       assert.strictEqual(actual, expected);
@@ -135,7 +131,7 @@ describe('replaceAll', function () {
 
     specify('should support empty searchValue', function () {
       const value = 'xxx';
-      const actual = replaceAllPolyfill('', '_', value);
+      const actual = replaceAllPonyfill('', '_', value);
       const expected = '_x_x_x_';
 
       assert.strictEqual(actual, expected);
@@ -143,7 +139,7 @@ describe('replaceAll', function () {
 
     specify('should replace all matches', function () {
       const value = 'ab cd ab cd ab cd';
-      const actual = replaceAllPolyfill('ab', 'ef', value);
+      const actual = replaceAllPonyfill('ab', 'ef', value);
       const expected = 'ef cd ef cd ef cd';
 
       assert.strictEqual(actual, expected);
@@ -151,7 +147,7 @@ describe('replaceAll', function () {
 
     context('given empty string', function () {
       specify('should return original value', function () {
-        assert.strictEqual(replaceAllPolyfill('a', 'c', ''), '');
+        assert.strictEqual(replaceAllPonyfill('a', 'c', ''), '');
       });
     });
 
@@ -159,7 +155,7 @@ describe('replaceAll', function () {
       const value = new String('ab cd ab cd ab cd');
       const searchValue = new String('ab');
       const replaceValue = new String('ef');
-      const actual = replaceAllPolyfill(searchValue, replaceValue, value);
+      const actual = replaceAllPonyfill(searchValue, replaceValue, value);
       const expected = 'ef cd ef cd ef cd';
 
       assert.strictEqual(actual, expected);
@@ -168,16 +164,16 @@ describe('replaceAll', function () {
     specify('should replace in very big strings', function () {
       const bigString = 'ab cd'.repeat(10000);
       const expected = 'ab ef'.repeat(10000);
-      const actual = replaceAllPolyfill('cd', 'ef', bigString);
+      const actual = replaceAllPonyfill('cd', 'ef', bigString);
 
       assert.strictEqual(actual, expected);
     });
 
     specify('should be curried', function () {
-      assert.strictEqual(replaceAllPolyfill('a', 'c', 'aba'), 'cbc');
-      assert.strictEqual(replaceAllPolyfill('a')('c', 'aba'), 'cbc');
-      assert.strictEqual(replaceAllPolyfill('a', 'c')('aba'), 'cbc');
-      assert.strictEqual(replaceAllPolyfill('a')('c')('aba'), 'cbc');
+      assert.strictEqual(replaceAllPonyfill('a', 'c', 'aba'), 'cbc');
+      assert.strictEqual(replaceAllPonyfill('a')('c', 'aba'), 'cbc');
+      assert.strictEqual(replaceAllPonyfill('a', 'c')('aba'), 'cbc');
+      assert.strictEqual(replaceAllPonyfill('a')('c')('aba'), 'cbc');
     });
   });
 });