+++ webpack.config-test.web.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const webpack = require('webpack'); // eslint-disable-line import/no-extraneous-dependencies
 const glob = require('glob'); // eslint-disable-line import/no-extraneous-dependencies
 const path = require('path');
 
@@ -7,7 +8,8 @@ module.exports = {
   mode: 'production',
   target: 'web',
   entry: [
-    '@babel/polyfill',
+    'core-js/stable',
+    'regenerator-runtime/runtime',
     ...glob.sync('./test/*.js', {
       ignore: ['./test/mocha-bootstrap.js'],
     }),
@@ -16,28 +18,28 @@ module.exports = {
     path: path.resolve('.'),
     filename: 'tmp-test-bundle.js',
   },
+  performance: {
+    maxEntrypointSize: 2000000,
+    maxAssetSize: 2000000,
+  },
+  resolve: {
+    fallback: {
+      assert: require.resolve('assert'),
+      process: require.resolve('process/browser'),
+    },
+  },
+  plugins: [
+    new webpack.ProvidePlugin({
+      process: 'process/browser',
+    }),
+  ],
   module: {
     rules: [
       {
         test: /\.(js)$/,
         loader: 'babel-loader',
         options: {
-          presets: [
-            [
-              '@babel/preset-env',
-              {
-                forceAllTransforms: true,
-              },
-            ],
-          ],
-          plugins: [
-            [
-              '@babel/plugin-transform-modules-commonjs',
-              {
-                loose: true,
-              },
-            ],
-          ],
+          presets: ['@babel/preset-env'],
         },
       },
     ],